% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_symbols.R
\name{update_symbols}
\alias{update_symbols}
\title{Update last HGNC correspondence database}
\usage{
update_symbols(url=NULL)
}
\arguments{
\item{url}{user can provide url (default is NULL)}
}
\value{
returns a data.frame containing gene symbols with status, previous symbols and synonyms as well as their corresponding entrezIDs
}
\description{
This function is used to update gene symbol correspondence from HGNC database
}
\examples{

\donttest{
HGNC <- update_symbols()
}
}
